requirements.txt
streamlit>=1.30
anthropic
python-dotenv
pytest>=7.0